sync latest mcx update on 04-07-2016, reduce global mem read

+++ src/mcx_utils.h
@@ -108,7 +108,7 @@ typedef struct MCXConfig{
         float unitinmm;     /*defines the length unit in mm for grid*/
         FILE *flog;         /*stream handle to print log information*/
         History his;        /*header info of the history file*/
-	float energytot, energyabs, energyesc;
+	double energytot, energyabs, energyesc;
         char rootpath[MAX_PATH_LENGTH];
         char kernelfile[MAX_SESSION_LENGTH];
 	char compileropt[MAX_PATH_LENGTH];
